backend/requirements.txt
fastapi
uvicorn
websockets
google-generativeai
fpdf
python-multipart